added recruitee vouching system, fixed api access mask issue, made news failover for if the forum DB is unavailable & refactored Discord job to use multiple channels

+++ r3mus/Infrastructure/IdentityModels.cs
@@ -453,7 +453,7 @@ namespace r3mus.Models
                     var req = WebRequest.Create("https://api.eveonline.com");
                     req.Timeout = 2000;
                     HttpWebResponse response = (HttpWebResponse)req.GetResponse();
-                    return EveXml.CreateApiKey(Convert.ToInt32(ApiKey), VerificationCode).Init().AccessMask;
+                    return EveXml.CreateApiKey(Convert.ToInt64(ApiKey), VerificationCode).Init().AccessMask;
                 }
                 catch (Exception ex) { return 0; } }
         }
